requirements.txt
setuptools>=57.0.0
pyspark>=3.1.2
pytest
requests==2.31.0
pydantic
pydantic_settings
pyyaml
kaggle
pandas